+++ packages/bento-design-system/src/index.ts
@@ -47,7 +47,7 @@ export * from "./Navigation/Navigation";
 export * from "./NumberInput/NumberInput";
 export * from "./Placeholder/Placeholder";
 export * from "./Popover/Popover";
-export * from "./ProgressBar/createProgressBar";
+export * from "./ProgressBar/ProgressBar";
 export * from "./SearchBar/createSearchBar";
 export * from "./Slider/createSlider";
 export * from "./Stepper/createStepper";
